backend-django/requirements/dev.txt
-r base.txt
pytest-django
pytest-cov
pytest-mock
ruff
